# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "User Management API",
        "contact": {
            "name": "Yigitcan Balcin"
        },
        "version": "1.0"
    },
    "paths": {
        "/users": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Gets all users",
                "responses": {
                    "200": {
                        "description": "Returns UserDTO array",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/application.UserDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Returns err.Error",
                        "schema": {
                            "$ref": "#/definitions/err.Error"
                        }
                    },
                    "500": {
                        "description": "Returns err.Error",
                        "schema": {
                            "$ref": "#/definitions/err.Error"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Creates new user",
                "parameters": [
                    {
                        "description": "CreateNewUser body",
                        "name": "application.CreateUserRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/application.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns UserDTO",
                        "schema": {
                            "$ref": "#/definitions/application.UserDTO"
                        }
                    },
                    "400": {
                        "description": "Returns err.Error",
                        "schema": {
                            "$ref": "#/definitions/err.Error"
                        }
                    },
                    "500": {
                        "description": "Returns err.Error",
                        "schema": {
                            "$ref": "#/definitions/err.Error"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Gets user by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns UserDTO",
                        "schema": {
                            "$ref": "#/definitions/application.UserDTO"
                        }
                    },
                    "400": {
                        "description": "Returns err.Error",
                        "schema": {
                            "$ref": "#/definitions/err.Error"
                        }
                    },
                    "500": {
                        "description": "Returns err.Error",
                        "schema": {
                            "$ref": "#/definitions/err.Error"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Deletes user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Returns err.Error",
                        "schema": {
                            "$ref": "#/definitions/err.Error"
                        }
                    },
                    "500": {
                        "description": "Returns err.Error",
                        "schema": {
                            "$ref": "#/definitions/err.Error"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Updates new user",
                "parameters": [
                    {
                        "description": "UpdateUser body",
                        "name": "application.UpdateUserRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/application.UpdateUserRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns UserDTO",
                        "schema": {
                            "$ref": "#/definitions/application.UserDTO"
                        }
                    },
                    "400": {
                        "description": "Returns err.Error",
                        "schema": {
                            "$ref": "#/definitions/err.Error"
                        }
                    },
                    "500": {
                        "description": "Returns err.Error",
                        "schema": {
                            "$ref": "#/definitions/err.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "application.CreateUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "application.UpdateUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "application.UserDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "err.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "validations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/err.ValidationError"
                    }
                }
            }
        },
        "err.ValidationError": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}